Handle shadow with directional light. Change shadow_hit method type parameter for tmin in Triangle and Plane objects, in order to fit the GeometricObject one and be used thanks to polymorphism.

+++ GeometricObjects/Primitives/Plane.cpp
@@ -78,7 +78,7 @@ bool Plane::hit(const Ray& ray, double& tmin, ShadeRec& sr) const {
   return false;
 }
 
-bool Plane::shadow_hit(const Ray& ray, double& tmin) const {
+bool Plane::shadow_hit(const Ray& ray, float& tmin) const {
   if (!shadows)
     return false;
